requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
fpdf2==2.7.8